--- packages/create-typink/.prettierignore
@@ -1,26 +1,0 @@
-build/
-build-*/
-coverage/
-node_modules/
-tmp/
-.DS_Store
-.env.local
-.env.development.local
-.env.test.local
-.env.production.local
-.npmrc
-.yarn/*
-!.yarn/releases
-!.yarn/plugins
-.pnp.*
-cc-test-reporter
-lerna-debug.log*
-npm-debug.log*
-tsconfig.*buildinfo
-yarn-debug.log*
-yarn-error.log*
-package-lock.json
-.idea
-.vscode
-yarn.lock
-LICENSE
