--- packages/create-typink/.prettierr.cjs
@@ -1,10 +1,0 @@
-module.exports = {
-  semi: true,
-  trailingComma: 'all',
-  singleQuote: true,
-  jsxSingleQuote: true,
-  printWidth: 120,
-  tabWidth: 2,
-  bracketSameLine: true,
-  importOrder: ['.*react.*', '^@polkadot/(.*)$', '^@dedot/(.*)$', '<THIRD_PARTY_MODULES>', '^[./]'],
-};
